Move command building to a separate file
